Liquibase and BookController changes: - add swagger descriptions in BookController - migrate logs from BookController to BookServiceImpl - make init test for BookController - add a liquibase configuration

+++ src/main/java/pl/kurs/booktask/model/Book.java
@@ -18,7 +18,7 @@ public class Book implements Serializable {
 
      @Id
      @GeneratedValue(strategy = GenerationType.IDENTITY)
-     private Long id;
+     private int id;
 
      @Column(nullable = false)
      private String title;
@@ -27,7 +27,7 @@ public class Book implements Serializable {
      private String author;
 
      @Column(name = "published_year", nullable = false)
-     private Integer publishedYear;
+     private int publishedYear;
 
      @Column(nullable = false, unique = true)
      private String isbn;
